backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
requires-python = ">=3.14"
dependencies = [
    "authlib>=1.6.6",
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.124.0",
    "google-auth>=2.45.0",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "readability-lxml>=0.8.4.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "datamodel-code-generator>=0.51.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
]

[tool.isort]
profile = "black"

[tool.mypy]
explicit_package_bases = true

[tool.pytest.ini_options]
pythonpath = ["."]
